adding vsphere updates to non native
