+++ scenarios/openshift/vmware_node_scenarios.yml
@@ -1,10 +1,17 @@
-# yaml-language-server: $schema=../plugin.schema.json
-- id: <vmware-node-stop/vmware-node-start/vmware-node-reboot/vmware-node-terminate>
-  config:
-    name: <node_name>                   # Node on which scenario has to be injected; can set multiple names separated by comma                           
-    label_selector: <label_selector>    # When node_name is not specified, a node with matching label_selector is selected for node chaos scenario injection 
-    runs: 1                             # Number of times to inject each scenario under actions (will perform on same node each time)                                                           
-    instance_count: 1                   # Number of nodes to perform action/select that match the label selector                                             
-    timeout: 300                        # Duration to wait for completion of node scenario injection
-    verify_session: True                # Set to True if you want to verify the vSphere client session using certificates; else False
-    skip_openshift_checks: False        # Set to True if you don't want to wait for the status of the nodes to change on OpenShift before passing the scenario 
+node_scenarios:
+  - actions:
+    - node_reboot_scenario
+    node_name:
+    label_selector: node-role.kubernetes.io/worker
+    instance_count: 1
+    timeout: 120
+    cloud_type: vmware
+  - actions:
+    - node_stop_start_scenario
+    node_name:
+    label_selector: node-role.kubernetes.io/worker
+    instance_count: 1
+    timeout: 360
+    duration: 10
+    cloud_type: vmware
+    parallel: false
